state: move methods to funcs
